+++ android/Lib/src/main/kotlin/com/darakeon/dfm/lib/ui/SwipeRefresh.kt
@@ -1,0 +1,32 @@
+package com.darakeon.dfm.lib.ui
+
+import android.content.Context
+import android.util.AttributeSet
+import android.widget.ListView
+import android.widget.ScrollView
+import android.widget.Toast
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+
+class SwipeRefresh(
+	context: Context,
+	attributes: AttributeSet
+) : SwipeRefreshLayout(context, attributes) {
+	var scrollChild: ScrollView? = null
+	var listChild: ListView? = null
+
+	override fun canChildScrollUp(): Boolean {
+		Toast.makeText(context, scrollChild.toString(), Toast.LENGTH_LONG).show()
+		Toast.makeText(context, listChild.toString(), Toast.LENGTH_LONG).show()
+
+		if (scrollChild != null) {
+			return scrollChild?.canScrollVertically(-1) ?: true
+		}
+
+		if (listChild != null) {
+			Toast.makeText(context, listChild?.firstVisiblePosition?.toString(), Toast.LENGTH_LONG).show()
+			return listChild?.firstVisiblePosition != 0
+		}
+
+		return true
+	}
+}