+++ docs/RELEASES.md
@@ -17,7 +17,7 @@ Legend:
 ## <a name="4.5.0.1"></a>4.5.0.1 :ant: <sup>`5`</sup>
 - [ ] Fix bug with android tests
 - [x] `201122>201122` Fix bug at future moves reports, showing zero value for move
-- [ ] Fix refresh gesture conflict with scroll of lists
+- [x] `201122>201123` Fix refresh gesture conflict with scroll of lists
 - [x] `201121>201122` Fix edit detailed move at site
 - [x] `201115>201120` Add create new task list at version changer (-q = quantity, -n = numbers)